--- Devantler.SOPSCLI.slnx
@@ -1,13 +1,0 @@
-<Solution>
-  <Configurations>
-    <Platform Name="Any CPU" />
-    <Platform Name="x64" />
-    <Platform Name="x86" />
-  </Configurations>
-  <Folder Name="/src/">
-    <Project Path="src/Devantler.SOPSCLI/Devantler.SOPSCLI.csproj" />
-  </Folder>
-  <Folder Name="/tests/">
-    <Project Path="tests/Devantler.SOPSCLI.Tests/Devantler.SOPSCLI.Tests.csproj" />
-  </Folder>
-</Solution>
