+++ DevantlerTech.SOPSCLI.slnx
@@ -1,0 +1,13 @@
+<Solution>
+  <Configurations>
+    <Platform Name="Any CPU" />
+    <Platform Name="x64" />
+    <Platform Name="x86" />
+  </Configurations>
+  <Folder Name="/src/">
+    <Project Path="src/DevantlerTech.SOPSCLI/DevantlerTech.SOPSCLI.csproj" />
+  </Folder>
+  <Folder Name="/tests/">
+    <Project Path="tests/DevantlerTech.SOPSCLI.Tests/DevantlerTech.SOPSCLI.Tests.csproj" />
+  </Folder>
+</Solution>
